refactor(iocuak): extract utilities from ValueServiceParameter module
