+++ packages/iocuak/src/e2e/definitions/common/models/worlds/ValueServiceWorld.ts
@@ -1,6 +1,6 @@
 import { IWorld } from '@cucumber/cucumber';
 
-import { ValueServiceParameter } from '../../parameters/ValueServiceParameter';
+import { ValueServiceParameter } from '../../parameters/valueService/ValueServiceParameter';
 
 export interface ValueServiceWorld extends IWorld {
   valueServiceParameter: ValueServiceParameter;
