+++ packages/iocuak/src/e2e/definitions/common/parameters/valueService/ValueServiceParameter.ts
@@ -1,0 +1,6 @@
+import { ValueBindingApi } from '../../../../../metadata/models/api/ValueBindingApi';
+
+export interface ValueServiceParameter {
+  binding: ValueBindingApi;
+  service: unknown;
+}
