+++ packages/iocuak/src/e2e/definitions/common/parameters/valueService/definition.ts
@@ -1,28 +1,8 @@
 import { defineParameterType } from '@cucumber/cucumber';
 import { IParameterTypeDefinition } from '@cucumber/cucumber/lib/support_code_library_builder/types';
 
-import { BindingTypeApi } from '../../../../metadata/models/api/BindingTypeApi';
-import { ValueBindingApi } from '../../../../metadata/models/api/ValueBindingApi';
-
-export interface ValueServiceParameter {
-  binding: ValueBindingApi;
-  service: unknown;
-}
-
-function getValueServiceParameter(): ValueServiceParameter {
-  const value: unknown = Symbol('value');
-
-  const binding: ValueBindingApi = {
-    bindingType: BindingTypeApi.value,
-    id: Symbol('value-service'),
-    value,
-  };
-
-  return {
-    binding,
-    service: value,
-  };
-}
+import { getValueServiceParameter } from './getValueServiceParameter';
+import { ValueServiceParameter } from './ValueServiceParameter';
 
 function valueServiceParameterDefinitionTransformer(): ValueServiceParameter {
   return getValueServiceParameter();
